ski-ascii.py converts Leica SKI ASCII format file to DynaML
Hi Craig,

This code was originally written by Nuddin Tengku, however I noted when using the converter, that the VCV Matrix calculation was incorrect, and giving a factor of ~100x different to the conversion performed in Zed, which does an identical conversion - causing big adjustment errors.

It is true that Zed could be used, but the value of the Python script however, lets us get away from a program based approach for Leica users, and very useful for those users who are not able to load software to their machine without administration rights.

I had also seen anomaly, where the final observation in the SKI Ascii file had been left out of the resultant measurement file from the original script. I altered the "Add baseline...." operation to just before the "return msrroot" operation at line 142 and this allowed all measurements to appear in the resultant msr.xml file. This is similar in placement to the "return stnroot" operation.

I also removed the "if linebatch >= 1 and len(linebatchdict[linebatch - 1]) == 10:" section of the  "Add baseline...." code, this may need to be checked further as to whether it can be validly excluded?

The last code change made, was to line 153 and 155, allowing and _underscore to be added before the stn and msr identifiers, this is just a "nice" thing to have in the file naming, in order for the extension not to get too mixed up in the source ASCII name.

Thanks,